requirements.txt
grpcio
grpcio-tools
protobuf
grpcio-reflection
windows-curses